Refactor route journeys into a separate component.
