+++ src/components/MergedJourneys.js
@@ -4,21 +4,12 @@ import compact from "lodash/compact";
 
 // Use memoization to prevent the combined journeys from updating too often.
 
-const MergedJourneys = ({
-  children,
-  routeJourneys,
-  areaJourneys = [],
-  selectedJourney,
-}) => {
+const MergedJourneys = ({children, areaJourneys = [], selectedJourney}) => {
   const currentJourneys = useMemo(() => {
     return mergeJourneys(compact([selectedJourney, ...areaJourneys]));
   }, [areaJourneys, selectedJourney]);
 
-  const routeAndSelected = useMemo(() => {
-    return currentJourneys.concat(routeJourneys);
-  }, [currentJourneys, routeJourneys]);
-
-  return children({currentJourneys, routeAndSelected});
+  return children({currentJourneys});
 };
 
 export default MergedJourneys;
