+++ src/components/map/MapEvents.js
@@ -7,7 +7,6 @@ import SelectedJourneyEvents from "../SelectedJourneyEvents";
 import MergedJourneys from "../MergedJourneys";
 import UnsignedVehicleEvents from "../UnsignedVehicleEvents";
 import {inject} from "../../helpers/inject";
-import RouteJourneys from "../RouteJourneys";
 
 const decorate = flow(observer, inject("state"));
 
@@ -31,40 +30,33 @@ const MapEvents = decorate(({children, state}) => {
         }
 
         return (
-          <RouteJourneys>
-            {({routeJourneys, loading: routeEventsLoading = false}) => (
-              <SelectedJourneyEvents>
-                {({
-                  journey: selectedJourney = null,
-                  loading: selectedJourneyLoading = false,
-                }) => (
-                  <UnsignedVehicleEvents>
-                    {({unsignedEvents = [], loading: unsignedEventsLoading = false}) => (
-                      <MergedJourneys
-                        routeJourneys={routeJourneys}
-                        areaJourneys={areaJourneys}
-                        selectedJourney={selectedJourney}>
-                        {({currentJourneys = [], routeAndSelected = []}) =>
-                          children({
-                            selectedJourney,
-                            routeJourneys,
-                            areaJourneys,
-                            unsignedEvents,
-                            currentJourneys,
-                            routeAndSelected,
-                            areaJourneysLoading,
-                            unsignedEventsLoading,
-                            routeEventsLoading,
-                            selectedJourneyLoading: selectedJourneyLoading,
-                          })
-                        }
-                      </MergedJourneys>
-                    )}
-                  </UnsignedVehicleEvents>
+          <SelectedJourneyEvents>
+            {({
+              journey: selectedJourney = null,
+              loading: selectedJourneyLoading = false,
+            }) => (
+              <UnsignedVehicleEvents>
+                {({unsignedEvents = [], loading: unsignedEventsLoading = false}) => (
+                  <MergedJourneys
+                    areaJourneys={areaJourneys}
+                    selectedJourney={selectedJourney}>
+                    {({currentJourneys = [], routeAndSelected = []}) =>
+                      children({
+                        selectedJourney,
+                        areaJourneys,
+                        unsignedEvents,
+                        currentJourneys,
+                        routeAndSelected,
+                        areaJourneysLoading,
+                        unsignedEventsLoading,
+                        selectedJourneyLoading: selectedJourneyLoading,
+                      })
+                    }
+                  </MergedJourneys>
                 )}
-              </SelectedJourneyEvents>
+              </UnsignedVehicleEvents>
             )}
-          </RouteJourneys>
+          </SelectedJourneyEvents>
         );
       }}
     </AreaJourneys>
