+++ src/components/map/UnsignedEventsLayer.js
@@ -4,6 +4,7 @@ import {observer} from "mobx-react-lite";
 import {inject} from "../../helpers/inject";
 import HfpMarkerLayer from "./HfpMarkerLayer";
 import JourneyLayer from "./JourneyLayer";
+import {useCurrentEvent} from "../../hooks/useCurrentEvent";
 
 const decorate = flow(observer, inject("state"));
 
@@ -11,25 +12,7 @@ const UnsignedEventsLayer = decorate(({unsignedEvents, state}) => {
   const {unixTime} = state;
   const prevEvent = useRef(null);
 
-  const currentEvent = useMemo(() => {
-    let event = null;
-    let currentTimeDiff = 60;
-
-    for (const evt of unsignedEvents) {
-      const timeDiff = Math.abs(unixTime - evt.recordedAtUnix);
-
-      if (timeDiff < currentTimeDiff) {
-        event = evt;
-        currentTimeDiff = timeDiff;
-      }
-
-      if (currentTimeDiff < 5) {
-        break;
-      }
-    }
-
-    return event;
-  }, [unsignedEvents, unixTime]);
+  const currentEvent = useCurrentEvent(unsignedEvents, unixTime);
 
   if (currentEvent) {
     prevEvent.current = currentEvent;
