+++ src/components/sidepanel/RouteDepartures.js
@@ -1,4 +1,4 @@
-import React, {useCallback} from "react";
+import React, {useCallback, useMemo} from "react";
 import {observer} from "mobx-react-lite";
 import getJourneyId from "../../helpers/getJourneyId";
 import styled from "styled-components";
@@ -6,9 +6,92 @@ import {Text} from "../../helpers/text";
 import SidepanelList from "./SidepanelList";
 import flow from "lodash/flow";
 import {inject} from "../../helpers/inject";
-import RouteDeparturesQuery from "../../queries/RouteDeparturesQuery";
 import EmptyView from "../EmptyView";
 import RouteDepartureItem from "./RouteDepartureItem";
+import {useQueryData} from "../../hooks/useQueryData";
+import gql from "graphql-tag";
+import {CancellationFieldsFragment} from "../../queries/CancellationFieldsFragment";
+
+export const routeJourneysQuery = gql`
+  query journeysByDateQuery(
+    $routeId: String!
+    $direction: Direction!
+    $date: Date!
+    $stopId: String!
+  ) {
+    routeDepartures(
+      routeId: $routeId
+      direction: $direction
+      date: $date
+      stopId: $stopId
+    ) {
+      id
+      index
+      isNextDay
+      isTimingStop
+      dayType
+      departureId
+      departureDate
+      departureTime
+      equipmentColor
+      equipmentType
+      extraDeparture
+      operatorId
+      terminalTime
+      recoveryTime
+      routeId
+      direction
+      stopId
+      isCancelled
+      cancellations {
+        ...CancellationFieldsFragment
+      }
+      journey {
+        id
+        journeyType
+        departureDate
+        departureTime
+        direction
+        routeId
+        originStopId
+        uniqueVehicleId
+        mode
+        _numInstance
+      }
+      plannedArrivalTime {
+        id
+        arrivalDate
+        arrivalDateTime
+        arrivalTime
+        isNextDay
+      }
+      observedArrivalTime {
+        id
+        arrivalDate
+        arrivalDateTime
+        arrivalTime
+        arrivalTimeDifference
+        loc
+      }
+      plannedDepartureTime {
+        id
+        departureDate
+        departureDateTime
+        departureTime
+        isNextDay
+      }
+      observedDepartureTime {
+        id
+        departureDate
+        departureDateTime
+        departureTime
+        departureTimeDifference
+        loc
+      }
+    }
+  }
+  ${CancellationFieldsFragment}
+`;
 
 const HeaderRowLeft = styled.span`
   display: block;
@@ -51,44 +134,66 @@ const RouteDepartures = decorate(({state, Time, Journey, UI}) => {
     Journey.setSelectedJourney(journeyToSelect);
   }, []);
 
+  const {routeId, direction, originStopId} = route;
+  const shouldSkip = !route || !routeId || !direction || !originStopId;
+
+  const queryVars = useMemo(
+    () => ({
+      routeId: routeId,
+      direction: parseInt(direction, 10),
+      stopId: originStopId || "",
+      date,
+    }),
+    [routeId, direction, originStopId, date]
+  );
+
+  const {data: departuresData = [], loading, error} = useQueryData(
+    routeJourneysQuery,
+    {
+      skip: shouldSkip,
+      variables: queryVars,
+    },
+    "route departures"
+  );
+
+  const departures = departuresData || [];
+
   return (
-    <RouteDeparturesQuery route={route} date={date}>
-      {({departures, loading, error, skipped}) => {
-        return error || (!loading && !skipped && departures.length === 0) ? (
-          <EmptyView
-            error={
-              error ? error : {emptyDataError: {message: "No data returned from server."}}
-            }
-          />
-        ) : (
-          <SidepanelList
-            testIdPrefix="journey"
-            focusKey={selectedJourneyId}
-            loading={loading}
-            header={
-              <JourneyListHeader data-testid="journey-list-header">
-                <HeaderRowLeft>
-                  <Text>filterpanel.planned_start_time</Text>
-                </HeaderRowLeft>
-                <span>
-                  <Text>filterpanel.real_start_time</Text>
-                </span>
-              </JourneyListHeader>
-            }>
-            {(scrollRef) =>
-              departures.map((departure) => (
-                <RouteDepartureItem
-                  key={departure.id}
-                  departure={departure}
-                  scrollRef={scrollRef}
-                  selectJourney={selectJourney}
-                />
-              ))
-            }
-          </SidepanelList>
-        );
-      }}
-    </RouteDeparturesQuery>
+    <>
+      {error || (!loading && !shouldSkip && departures.length === 0) ? (
+        <EmptyView
+          error={
+            error ? error : {emptyDataError: {message: "No data returned from server."}}
+          }
+        />
+      ) : (
+        <SidepanelList
+          testIdPrefix="journey"
+          focusKey={selectedJourneyId}
+          loading={loading}
+          header={
+            <JourneyListHeader data-testid="journey-list-header">
+              <HeaderRowLeft>
+                <Text>filterpanel.planned_start_time</Text>
+              </HeaderRowLeft>
+              <span>
+                <Text>filterpanel.real_start_time</Text>
+              </span>
+            </JourneyListHeader>
+          }>
+          {(scrollRef) =>
+            departures.map((departure) => (
+              <RouteDepartureItem
+                key={departure.id}
+                departure={departure}
+                scrollRef={scrollRef}
+                selectJourney={selectJourney}
+              />
+            ))
+          }
+        </SidepanelList>
+      )}
+    </>
   );
 });
